Move GetRoute.Model, PackageSchema to own file, PackageSchema to API level
